Optimised today's solution
It used to run in 28 seconds, but only when printing both solutions; if
I print either partOne or partTwo, it was just about two seconds. Some
discussion on #haskell revealed that the global 'sequence*' lists might
be the cause, and indeed folding them helped.

This new version now runs in 2.8 seconds.